Document the benchmark baseline
A full local run backs the docs: SqlBound stays within 1-9% of
hand-written ADO.NET on point operations, allocates exactly what the
hand-written loop allocates on 1000-row lists, and beats Dapper in
every category on both time and allocations. The row type switches to
SQLite''s natural provider types (long/double) because Dapper''s
constructor mapping requires exact type matches and crashed on
narrowing - noted in the docs as a fairness decision so all three
contenders do identical work. The generated null guards cost 20% on
list time versus raw; documented as the deliberate trade for
descriptive NULL errors.

+++ bench/SqlBound.Benchmarks/BenchmarkQueries.cs
@@ -21,4 +21,6 @@ public static partial class BenchmarkQueries
         DbConnection connection, int id, decimal price, CancellationToken cancellationToken = default);
 }
 
-public sealed record Item(int Id, string Name, decimal? Price);
+// Uses SQLite's natural provider types (INTEGER => long, REAL => double) so all three contenders
+// do identical work: Dapper's constructor mapping requires exact type matches and cannot narrow.
+public sealed record Item(long Id, string Name, double? Price);
